feat(coding-agent): add settings toggle and /code-index slash command
- Add tools.codeSearchEnabled setting (default: false) to settings schema
  under Tools → Available Tools; code_search tool is now gated by this
  setting instead of the OMP_ZVEC_ENABLED env var
- Add /code-index slash command to explicitly build or rebuild the Zvec
  code search index; uses loadMnemopi() for lazy module loading
- Refactor indexing logic into standalone indexWorkspace() function in
  code-search-index.ts, shared by both the slash command and the tool
- Tool no longer auto-indexes on first search; only searches the existing
  index and tells users to run /code-index when the index is empty
- Update tool description to mention /code-index requirement
