+++ packages/coding-agent/CHANGELOG.md
@@ -13,6 +13,8 @@
 - Added `/queue <message>` plus `->` / `=>` composer shorthand for follow-up messages that wait until the agent yields. The shorthand opens a dim `Queueing` header and splits sequential numeric, Roman-numeral, or alphabetic lists into separately highlighted queue entries.
 - Added per-model TPS/TTFT tracking: every completed assistant turn folds its timing into recency-weighted aggregates in `~/.omp/agent.db`, and the /models browser shows measured speed — a right-aligned `118t/s` column on wide terminals (plus TTFT, e.g. `0.9s 118t/s`, when wider) and `~118t/s · 0.9s ttft` facts in the selection detail line — with no dependency on the `omp stats` session scan.
 - Added `code_search` tool for semantic code search using Zvec in-process vector database with HNSW indexing and BM25 full-text search; indexes the workspace on first use and supports hybrid (vector + FTS) search when embeddings are available, falling back to FTS-only mode otherwise.
+- Added `tools.codeSearchEnabled` setting (default: off) to enable/disable semantic code search in the /settings panel.
+- Added `/code-index` slash command to explicitly build or rebuild the Zvec code search index; the `code_search` tool no longer auto-indexes on first use.
 
 ### Changed
