+++ packages/coding-agent/test/tool-discovery/initial-tools.test.ts
@@ -27,6 +27,7 @@ const allToolsSettings = Settings.isolated({
 	"browser.enabled": true,
 	"checkpoint.enabled": true,
 	"todo.enabled": true,
+	"tools.codeSearchEnabled": true,
 	"memory.backend": "mnemopi",
 	"autolearn.enabled": true,
 	"tools.discoveryMode": "all",
